requirements.txt
tensorflow
streamlit
numpy
pillow
streamlit-extras